Refactor: Controller 구현
Controller를 구현하여 다른 Controller들이 해당 Controller를 상속받게 함
